backend/requirements.txt
fastapi[standard]
bcrypt==3.2.2
python-jose
passlib
websockets
pydantic
sqlalchemy